Add `custom_params` to `rsyslog::client` for custom templates
When doing lookups in templates, variables that are not explicitly
scoped with `scope.lookupvar()` must be in the scope of the `template()`
function call. `rsyslog::client` allows custom templates to be called,
but doesn't allow custom parameters to be passed for the alterate
template to use.

This pull request adds a `custom_params` parameter so a hash of extra
parameters may be passed for the custom template to access without
explicit scoping knowledge.