Add parameter 'v8plus' for rsyslog version 8+
There has been some significant config changes in versions 8+ of
Rsyslog, one of which is causing deprecated warnings with the Puppet
module. This commit adds are class boolean parameter 'v8plus' to signify
if the version of Rsyslog is version 8 or greater.

Also, the deprecation warning is fixed (correct syntax is used).