Allow option to avoid loading imfile module twice.
This throws a warning with version 8 of rsyslog.

The best way to use imfile is to define the module list in
extra_modules. However, if you still try to create an imfile without
adding that it will still work the old way (adding modload imfile to the
begining of the imfile).