Update database.conf.erb
Added an rsyslog template for sql formatting, which properly separates the Syslogtag and ProcessID into their respective fields when inserting into an SQL database.

When using the Adiscon Log Analyzer, this not only allows proper filtering on the Syslogtag field, but also corrects the statistics graph generated from that field. This also facilitates searching and filtering by ProcessID field.