Contain install, config, and service classes
Prior to this, it was impossible to make a relationship to the 'rsyslog'
class from another class and expect the install, config, and service
classes to honor the relationship. The reason being that those classes
were not contained.

This fixes that by using the contain function to properly contain the
install, config, and service classes within the rsyslog class.

https://docs.puppet.com/puppet/4.9/lang_containment.html#the-contain-function

This lets someone make the following relationship, for example when
needing to make sure that the EPEL YUM repos are managed before
installing rsyslog:

    include epel
    include rsyslog::client

    Class['epel'] -> Class['rsyslog']

Note that the relationship must be made to the 'rsyslog' class when
using just 'rsyslog::client'.

Also note that the contain function was introduced in Puppet 3.4. As
such, the version of Puppet required to use this module has been
increased to 3.4.