Add lookup_tables, custom confdir, custom targets, and managed_service options (#20)
* Add lookup_table JSON file features

* New lookup_table_target (String) parameter for specifying the path
to the JSON file to be generated.
* Added lookup_tables (Optional[Hash]) parameter to the client, server,
and config classes.
* Created new rsyslog::component::lookup_table defined type to generate
the JSON files.
* Created new rsyslog::config::lookup_tables class to parse the
lookup_tables parameter and pass that data to the defined type.
* Added new unit tests for new feature

TODO:
* Add lookup_table() and lookup() RainerScript config generation

* Update README and remove unused parameter

* Fix failing travis tests

Updated travis.yml file to test against supported ruby+puppet
combinations.

Ruby 1.9.3 is not supported in any version of puppet 4 and some tests
are failing to run purely on a requirement of ruby 2.x. Puppet 4 is
built against Ruby 2.1.9, so all tests for puppet 4 have been migrated
to Ruby 2.1.9.

Add tests for Puppet 5 + ruby 2.4.1

* Update dependency matrix and tests

Remove tests for Puppet 4.3.0 and earlier. Test failures appear to
indicate that module functionality uses features that are not available
in puppet 4.3.0 and earlier.

Increased upper limit on the puppetlabs/concat support to < 5.0.0. As
the tests are already testing against concat master, we know it's
compatible with concat 4.x.

Added Puppet 5.x support to the metadata.json file. Been testing using
Puppet 5.0.1 since I started on this PR.

* Add the ability to manage the service or not

* Refactor components and config to be more dynamic

Removed the concat from the config class and created a new
defined type called generate_concat. This defined type will take in
2 parameters - confdir and target - and create a concat resource off
those values if the resource does not already exist.

The component defines have been refactored to use this new defined type
and the config classes will set the default for confdir to $::rsyslog::confdir
if no other directory is specified.

* Add confdir refactor changes to lookup_tables

* Add default data for manage_service and set to true

* Convert global flattened config into a default expression rather than a dynamic hash map

* Update Documentation

* Add rsyslog::lookup_tables_priority parameter and defaults

* Add priority default

* Fix bug where concat fragment always used default confdir

* Update main_queue to allow for custom confdir and target files